requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
python-dotenv==1.2.2
anthropic==0.84.0
openai==2.26.0
Pillow==12.1.1
